% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/todor.R
\name{build_markdown_report}
\alias{build_markdown_report}
\title{Build TODO report in markdown syntax}
\usage{
build_markdown_report(markers)
}
\arguments{
\item{markers}{List of todor markers.}
}
\description{
Extracts the list of unique files which contain a \code{todor}
  marker and applies \code{extract_markers_to_md} to each of these files.
}
