% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{build_rstudio_markers}
\alias{build_rstudio_markers}
\title{Build Rstudio Markers}
\usage{
build_rstudio_markers(markers)
}
\arguments{
\item{markers}{list of markers}
}
\description{
Build Rstudio Markers
}
