% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{clean_comments}
\alias{clean_comments}
\title{Clean line from comment tags}
\usage{
clean_comments(line)
}
\arguments{
\item{line}{character with comment tag to remove}
}
\value{
cleaned character
}
\description{
Clean line from comment tags
}
\examples{
clean_comments("#' TODO abc abc") #"TODO abc abc"
}
