% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{create_markers}
\alias{create_markers}
\title{Create markers}
\usage{
create_markers(todo.list)
}
\arguments{
\item{todo.list}{list of files with lists of items detected in it}
}
\value{
markers list
}
\description{
Create markers
}
