% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/todor.R
\docType{data}
\name{DEFAULT_PATTERNS}
\alias{DEFAULT_PATTERNS}
\title{Default patterns}
\format{An object of class \code{character} of length 1.}
\usage{
DEFAULT_PATTERNS
}
\description{
Default patterns
}
\keyword{datasets}
