% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{extract_markers_to_md}
\alias{extract_markers_to_md}
\title{Extract markers to markdown}
\usage{
extract_markers_to_md(file, markers)
}
\arguments{
\item{file}{Name of file. Used to extract TODOs in that file from the list of
markers.}

\item{markers}{List of \code{todor} markers.}
}
\description{
Extracts all \code{todor} markers in a given file and converts
  them to bullet-pointed markdown syntax. The file name is printed in bold at
  the top of each section.
}
