% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{find_package}
\alias{find_package}
\title{Find package}
\usage{
find_package(path = getwd())
}
\arguments{
\item{path}{character with path to directory with R code files}
}
\value{
package path
}
\description{
Find package
}
