% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{find_pattern}
\alias{find_pattern}
\title{Find pattern}
\usage{
find_pattern(text, patterns = c("TODO", "FIXME"))
}
\arguments{
\item{text}{character with text}

\item{patterns}{character vector}
}
\value{
character with pattern in brackets or NULL
}
\description{
Find patterns like
\code{
# TODO text
#TODO Test this thing.
#TODO: Test this thing.
#TODO - Test this thing.
<!-- TODO Test this thing. -->
<!-- TODO: Test this thing. -->
<!-- TODO - Test this thing. -->
}
}
