% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{is_multiline}
\alias{is_multiline}
\title{Find multiline patterns}
\usage{
is_multiline(text)
}
\arguments{
\item{text}{character with text}
}
\value{
boolean
}
\description{
Internal function used to detect if a pattern is followed by a commented non-epmty line
}
