% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{list_files_with_extension}
\alias{list_files_with_extension}
\title{List files with given extension}
\usage{
list_files_with_extension(extension, search_path)
}
\arguments{
\item{extension}{character with extension}

\item{search_path}{character with path to start searching from}
}
\value{
list of files with specified extension
}
\description{
It lists recursively with full path names.
}
