% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{process_file}
\alias{process_file}
\title{Process file}
\usage{
process_file(filepath, patterns)
}
\arguments{
\item{filepath}{character with}

\item{patterns}{vector of characters with given patterns to detect,
e.g. c("TODO", "BUG")}
}
\value{
list of markers (which are lists describing properties
of detected item from find_pattern)
}
\description{
It calls \code{find_pattern} on given file and return detected markers or NULL.
}
