% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{todolist_to_file}
\alias{todolist_to_file}
\title{Write detected markers to file}
\usage{
todolist_to_file(todo.list, file_output)
}
\arguments{
\item{todo.list}{named output from \code{process_file}. list of files with lists of items detected in it}

\item{file_output}{character string with file to write to. File name can be provided with .txt extension or without. Any other extension will be replaced.}
}
\value{
output of formatted markers into file
}
\description{
Takes the output from \code{process_file} (named with file names) and writtes it to a file.
Matched patterns will be grouped by file and type (e.g. TODO, BUG, etc.)
}
