% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/todor.R
\name{todor_file}
\alias{todor_file}
\title{Todor file}
\usage{
todor_file(file_name, todo_types = NULL, file_output = NULL)
}
\arguments{
\item{file_name}{character with file name}

\item{todo_types}{vector with character describing types of elements to detect.
If NULL default items will be used.}

\item{file_output}{character string with file to write to. If NULL, matches appear as RStudio markers.}
}
\description{
Todor file
}
