% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/todor.R
\name{todor_package}
\alias{todor_package}
\title{Todor Package addin}
\usage{
todor_package(todo_types = NULL)
}
\arguments{
\item{todo_types}{vector with character describing types of elements to detect.
If NULL default items will be used.}
}
\description{
Called on packages. Checks all places in the code which require amendents
as specified in \code{todo_types}.
It triggers rstudio markers to appear.
}
