% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/todor.R
\name{todor}
\alias{todor}
\title{TODOr}
\usage{
todor(todo_types = NULL, search_path = getwd(), file = NULL)
}
\arguments{
\item{todo_types}{vector with character describing types of elements to detect.
If NULL, the default items will be used.}

\item{search_path}{vector with paths that contains comments you are looking for.}

\item{file}{character with path to file. If not NULL the search_path will be ignored.}
}
\description{
This package helps you to find all rows in your code with places
to be fixed / improved in the future.

It checks all places in the code for amendent comments on R and r files
(eg. TODO, FIXME etc.) as specified in \code{todo_types}.
It triggers rstudio markers to appear.
}
\details{
When option \code{todor_rmd} is set to TRUE it searches also through Rmd files.
Unless option \code{todor_exlude_packrat} is set to FALSE, all files in the packrat directory
are excluded.

This should be called on project that are not R packages. Check \code{todor_package}
to search only through the package folders.
}
\examples{
\donttest{
todor()
}
}
