% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/todor.R
\name{todor}
\alias{todor}
\title{TODOR
This package helps you to find all places in your code with places
to be filled in the future.}
\description{
TODOR
This package helps you to find all places in your code with places
to be filled in the future.
}
